Fix review comment, raise navbar and separate from sidebar
Since the sidebar can slide out on small screens it's better to let it
be a separate element instead of melting into the navbar.  That's why
we also change the colors, taking the opportunity to create a white bg
for the navbar in light mode.

We also adjust the starting row/line for the sidebar so its shadow edge
does not overflow into the drop shadow of the navbar.

Finally, with these changese we can drop the extra <hr> separator that
only disturbed the flow of the sidebar text and main content.  Meaning
we have to add a bit of top margin before the sidebar text starts to
prevent it from being too close to the navbar.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>